Implemented and verified: Unit::isBeingConstructed, Unit::isConstructing, Unit::isMorphing, and Unit::isRepairing.
isBeingConstructed will return false if construction has been halted (such as if the SCV dies or halts construction to go do something else).

Also reimplemented Unit::getBuildUnit. When called on an SCV constructing a building, it will return the building it is creating (or null if it hasn't started construction yet). When called on a building that is Being Constructed, it will return the SCV that is constructing it (or null if construction has halted).